requirements/base.txt
Django==1.8.3

# Settings management
#django-configurations==0.9
git+https://github.com/jezdez/django-configurations.git@5ece107044

# database
psycopg2==2.6.1
dj-database-url==0.3.0
django-pgjson==0.3.1

# templates
django-jinja==1.4.1

# for bcrypt passwords
bcrypt==2.0.0

# Third party modules
django-extensions==1.5.5
djangorestframework==3.1.3
django-mptt==0.7.4
# GEO
djangorestframework-gis==0.9.3
geopy==1.10.0
django-recurrence==1.2.0
arrow==0.6.0
django-redis-sessions==0.5.0

# static file serving
whitenoise==2.0.2
